\name{Athletic}
\alias{Athletic}
\non_function{}

\title{Intercollegiate Athletic Program Data Set}

\description{
The \code{Athletic} data set has 168 observations and 7 variables.
The data come from a survey of stakeholders different (students, alumni, faculty, and
athletic department employees) of a large US state university. The variables in
the data set are conjoint analysis based relative importance weights for seven
potential indicators of intercollegiate athletic program success.
}

\format{
  This data set contains the following variables:
  \describe{
    \item{Win}{
      The importance of winning (won/loss record percentage) in determining the
      respondent's judgment of an intercollegiate athletic program's success. 
      }
    \item{Grad}{
      The importance of student athlete graduation rates in determining the
      respondent's judgment of an intercollegiate athletic program's success. 
      }
    \item{Violat}{
      The importance of NCAA rule violations in determining the
      respondent's judgment of an intercollegiate athletic program's success. 
      }
    \item{Attnd}{
      The importance of home game attendance in determining the
      respondent's judgment of an intercollegiate athletic program's success. 
      }
    \item{Fem}{
      The importance of gender equity (based on the ratio of female to male
      student athletes) in determining the
      respondent's judgment of an intercollegiate athletic program's success. 
      }
     \item{Teams}{
      The importance of the number of different sports teams in determining the
      respondent's judgment of an intercollegiate athletic program's success. 
      }
    \item{Finan}{
      The importance of the financial success of the program in determining the
      respondent's judgment of an intercollegiate athletic program's success. 
      }
    }
  }

\source{
 Wolfe, Richard A. and Daniel S. Putler (2002), "How Tight are the Ties that
 Bind Stakeholder Groups?, \emph{Organizaton Science}, 13(January-February),
 64-82.
}


\usage{
data(Athletic)
}

\keyword{datasets}
